Force generated app resources to be overwritten
Due to the way the underlying Ant Copy task is
implemented, it's possible for the file copy to
not be performed if Ant determines the destination
file is up to date with the source file. For the
generated test apps, we always want the latest and
greatest version of the files.