Fix issue with Spring Security Core debug filter
I pulled out the logic that recreates the entire
springSecurityFilterChainProxy bean in order to
reduce the potential for compatibility issues with
other plugins and applications that might depend
on the default configuration provided by the SSC
plugin. This ensures the plugin performs only the
dynamic modifications outlined in the documentation.