Update to Spring OAuth 2.0.6.RELEASE
Spring MVC 4 allows the HttpMethod to be passed into controller
methods. This will be null when using a version of Grails that
depends on Spring MVC 3.

We assume the requestMethod passed to the token endpoint is a
POST and default the value to such. We should probably revisit
this in the future if we want plugin consumers to have the
flexibility to specify the HTTP methods allowed for token requests.